updates readme
