+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "pnp-pdf",
-  "version": "0.0.3",
+  "version": "0.0.4",
   "main": "web/index.js",
   "types": "types/web/index.d.ts",
   "description": "Plug and play build of Mozilla's PDF.js library for React.js applications.",
