+++ web/react_pdf_viewer.d.ts
@@ -1,0 +1,12 @@
+import { CSSProperties } from "react";
+
+declare type PDFViewerProps = {
+  file: string;
+  style: CSSProperties;
+}
+
+declare const PDFViewer: (props: PDFViewerProps) => JSX.Element;
+
+export { PDFViewer };
+
+export default PDFViewer;
